create client-record and error handling

+++ routers/client.js
@@ -1,27 +1,28 @@
 const express = require('express');
 const {db} = require("../utils/db");
-
 const clientRouter = express.Router();
+const {ClientRecord} = require('../records/client-record')
+const {NotFoundError} = require("../utils/errors");
 
 clientRouter
     .get('/', (req, res) => {
         res.render('client/list-all.hbs', {
-            clients: db.getAll()
-        })
-    })
-    .get('/:id', (req, res) => {
-        res.render('client/one.hbs', {
-            client: db.getOne(req.params.id)
-        })
-    })
-    .post('/', async(req, res) => {
-        const id = await db.create(req.body);
-        console.log(id);
-        res.render('client/added.hbs', {
-            name: req.body.name,
-            id: id
+            clients: db.getAll(),
         });
     })
+
+    .get('/:id', (req, res) => {
+        const client = db.getOne(req.params.id);
+
+        if (!client) {
+            throw new NotFoundError()
+        }
+        res.render('client/one.hbs', {
+            client,
+        })
+
+    })
+
     .put('/:id', (req, res) => {
         const id = req.params.id;
         const update = db.update(req.params.id, req.body);
@@ -31,18 +32,43 @@ clientRouter
             id
         })
     })
-    .delete('/:id', (req, res) => {
-        db.delete(req.params.id);
-        res.render('client/delete.hbs')
+
+    .post('/', (req, res) => {
+        const id = db.create(req.body)
+        res
+            .status(201)
+            .render('client/added.hbs', {
+                id,
+            })
     })
+
+    .delete('/deleted/:id', (req, res) => {
+        const client = db.getOne(req.params.id);
+
+        if (!client) {
+            throw new NotFoundError()
+        } else {
+            db.delete(req.params.id);
+        }
+        res.render('client/deleted.hbs')
+    })
+
     .get('/form/add', (req, res) => {
         res.render('client/forms/add.hbs');
     })
+
     .get('/form/edit/:id', (req, res) => {
+        const client = db.getOne(req.params.id);
+
+        if (!client) {
+            throw new NotFoundError()
+        }
         res.render('client/forms/edit.hbs', {
-            client: db.getOne(req.params.id)
+            client,
         });
     })
+
+
 module.exports = {
-    clientRouter
+    clientRouter,
 }
